move related code to pandas module
