--- connector-agent-python/src/writers/mod.rs
@@ -1,1 +1,0 @@
-pub mod pandas;
